PIPELINE-4259: Rename publish CLI command to gap-events
Renames the PublishGaps CLI command to GapEvents (subcommand "publish" ->
"gap-events"), and renames the underlying pipelines/publish/ module to
pipelines/events/ with PublishGapsConfig -> GapEventsConfig, to match the
existing GapEventQuery/GapEventsTableConfig naming and pair semantically
with the raw-gaps command. Updates the job name literal, docs, and the
sample-publish.yaml -> sample-gap-events.yaml config file accordingly.
