+++ src/pipe_gaps/pipelines/events/table_config.py
